fix: Modify user movie score code and change score type from double to float

+++ src/main/java/org/stepup/cinesquareapis/movie/repository/MovieSimpleRepository.java
@@ -23,6 +23,5 @@ public interface MovieSimpleRepository extends JpaRepository<MovieSimple, Intege
     @Transactional
     @Modifying
     @Query("UPDATE MovieSimple m SET m.score = :score WHERE m.movieId = :movieId")
-    int updateAvgScore(Double score, Integer movieId);
-
+    int updateAvgScore(Float score, Integer movieId);
 }
